Add CH341 USB-to-SPI/GPIO driver with asynchronous USB transfers
This commit introduces a new driver for the CH341 USB-to-SPI/GPIO interface, utilizing ctypes to directly call libusb for asynchronous operations. The implementation includes support for SPI transfers, GPIO control, and proper handling of USB transfers with error management. Key features include:

- Singleton pattern for driver instance management.
- Asynchronous transfer handling with callbacks.
- GPIO pin manipulation methods for setting direction and state.
- SPI transfer methods with support for both output and input operations.
- Comprehensive error handling and logging throughout the driver.

This driver aims to provide a reliable interface for interacting with CH341 devices in Python.
