+++ pyproject.toml
@@ -38,6 +38,7 @@ hardware = [
     "python-periphery>=2.4.1",
     "spidev>=3.5",
     "pyserial>=3.5",
+    "pyusb>=1.2.1",
 ]
 websocket = [
     "websockets>=11.0.0",
